qemu and bochs handle drives differently, it seems.
Currently, qemu double faults when detecting drives on the secondary bus. The reasons are yet unknown, but the tests that pass on bochs don't work on qemu.
As a result, the code is currently unusable on qemu.

Apparently, detecting ATA drives is hard work. There's a lenghty discussion at http://f.osdev.org/viewtopic.php?f=1&t=20690, still waiting to be read and understood.

The commit tries to detect the absence of buses / controllers / drives. Works for bochs.