yet again, forgot to commit some files previously. ata.c and ata.h are about the bochs/qemu problem in last commit
follo_clusterchain() is able to find all the clusters related to a file. Next step, load the corresponding sectors from disk.

Also note that it seems there's a bug with ata_readblock, the ptr returned is not nulled, so it sometimes produces garbage. Fixing it is on the todo list.