fat_parse_path("\path\to\file") parses the directory path and uses fat_findcluster to return the starting cluster of the last file/dir in path.
fat_raw_to_humanreadable(table->file_name) provides us with human readable file names.

As always, the functions are not perfect and there might be cases where there be dragons. Some of the things on the todo ilist include:
* relative paths
* trailing \, as in \boot\grub\
* separating files from dirs, ISDIR, ISFILE or something like that

In plain words, we're now able to locate any file in the hard drive and thus continue with loading any file to memory. A memory manager is needed, obviously