kfree() and krealloc(). kfree works like charm. krealloc can decrease the allocated memory and handles situations where no changes are needed. Next up, handling increasing memory and relocating memory areas
In hindsight, I probably should have planned the memory manager together with paging. When paging will finally be initialized, some changes might be needed again. On the other hand, this is a learning experience and I like to solve problems as they come. This way, I will find out what's actually needed in osdev to create a functional system.