Paging
kmalloc() now allocates all memory in 4kb chunks. This probably causes a lot of internal fragmentation, but don't mind at this point. Also fixed some compiler warning in kmalloc

Paging itself is very basic, allocate a directory and a table to hold memory from 0 to 4Mb (is this identity paging, I'm not sure) and init paging. It is unknown if the code works or not. There is a triple fault after setting paging on in CR0, and it seems the exception handler doesn't work and all exceptions are double faults. Next up, proper exception handling.