This commit has been made while learning to use git rebase --interactive and squashes to meld commits together. Let's see how it goes, we've already made one mistake..
two things are done in this commit:

1. Fixed an interrupt bug where.. interrupts simply didn't work correctly. Now we're able to raise a page fault when a page fault occurs. Mostly the commit is cosmetic changes, though. Done in desctables, isr, interrupts.s boot.s and so forth

2. Paging.
Paging can be enabled without a triple fault, and a page fault handler works. This is the beginning of a beautiful friendship.
Next step: figure out what a (kernel) heap is, allocate pages for the user mode, learn to associate virtual addresses with physical ones, and load a user program.